Move Analytics into webpack land and stop using WebWorkers.
