+++ src/analytics/tagCloud.js
@@ -1,7 +1,4 @@
-onmessage = (e) => {
-  if (!e) return;
-  const cards = e.data;
-
+async function tagCloud(cards) {
   const tags = {};
   cards.forEach((card) =>
     card.tags.forEach((tag) => {
@@ -13,10 +10,12 @@ onmessage = (e) => {
     }),
   );
   const words = Object.keys(tags).map((key) => ({ value: key, count: tags[key] }));
-  postMessage({
+  return {
     type: 'cloud',
     description:
       'Tags in your cube with random colors weighted by the expected number of cards with that tag a player will open on average.',
     words,
-  });
-};
+  };
+}
+
+export default tagCloud;
