+++ src/analytics/typeBreakdownCount.js
@@ -1,33 +1,6 @@
-function GetColorCat(colors) {
-  if (colors.length === 0) {
-    return 'Colorless';
-  }
-  if (colors.length > 1) {
-    return 'Multi';
-  }
-  if (colors.length === 1) {
-    switch (colors[0]) {
-      case 'W':
-        return 'White';
-      case 'U':
-        return 'Blue';
-      case 'B':
-        return 'Black';
-      case 'R':
-        return 'Red';
-      case 'G':
-        return 'Green';
-      case 'C':
-      default:
-        return 'Colorless';
-    }
-  }
-}
+import { GetColorCat } from 'utils/AnalyticUtils';
 
-onmessage = (e) => {
-  if (!e) return;
-  const cards = e.data;
-
+async function typeBreakdownCount(cards) {
   const TypeByColor = Object.fromEntries(
     ['Creatures', 'Enchantments', 'Lands', 'Planeswalkers', 'Instants', 'Sorceries', 'Artifacts', 'Total'].map(
       (header) => [
@@ -96,7 +69,7 @@ onmessage = (e) => {
     }
   }
 
-  postMessage({
+  return {
     type: 'table',
     description: 'The count of cards in that type and color, percentages are relative to the bottom row of totals.',
     tables: [
@@ -124,5 +97,7 @@ onmessage = (e) => {
         ],
       },
     ],
-  });
-};
+  };
+}
+
+export default typeBreakdownCount;
