+++ src/components/analytics/Cloud.js
